fix(settings): Move change password to under security section in settings
