+++ packages/fxa-settings/src/components/Profile/en-US.ftl
@@ -5,12 +5,6 @@ profile-picture =
   .header = Picture
 profile-display-name =
   .header = Display name
-profile-password =
-  .header = Password
-# This is a string that shows when the user's password was created.
-# Variables:
-#   $date (String) - a localized date and time string
-profile-password-created-date = Created { $date }
 profile-primary-email =
   .header = Primary email
