+++ packages/fxa-settings/src/components/Security/en-US.ftl
@@ -1,3 +1,9 @@
 ## Security section of Setting
 
 security-heading = Security
+security-password =
+  .header = Password
+# This is a string that shows when the user's password was created.
+# Variables:
+#   $date (String) - a localized date and time string
+security-password-created-date = Created { $date }
